requirements.txt
google-cloud-vision
google-cloud-storage
scikit-learn
translators
jinja2
pillow
manga_ocr
